Updated SQL (MySQL still needs to be tested).
NOTE::: at this point, upgrading to 0.2 is pretty manual if you're already
running 0.1.x (if you don't care about the old data getting moved into the
new tables, this is a moot point).  To do it, simply update the site & lock
it (create a "upgrade.lock" file in the rw directory), then run the schema &
upgrade SQL files, then remove the lock file.  The site should do a very
minor update (creating version information for cs-webdblogger) and proceed as
normal.