Moved cs_versionAbstract to cs-content, abstract class into abstract folder.
NOTE::: as stated in a recent commit to cs-content, moving cs_versionAbstract
into the cs-content project means the two projects are not inextricably
attached to one another, and that this project is truly an extension of
cs-content.  Previously, since everything in cs-content required
cs_versionAbstract{}, this project would ALWAYS have to be available (which
wasn't a problem when cs_versionAbstract{} was in it's own project,
"cs-versionparse").

/cs_authToken.class.php:	
	* fix location of cs_webapplibsAbstract.
/cs_version.abstract.class.php [MOVED]:
	* moved to cs-content/abstract/
/cs_webdblogger.class.php:
	* fix location of cs_webapplibsAbstract.
/cs_webdbupgrade.class.php:
	* fix location of cs_webapplibsAbstract.
/tests/testOfCSWebAppLibs.php:
	* test_version_basics() [DELETED]
	* test_check_higher() [DELETED]
	* middleTestClass{} [DELETED]
/tests/files/version* [DELETED]:
	* moved to cs-content for testing cs_versionAbstract