First unit tests for cs_phpDB work.
I've done some very simple changes to get the tests to work.  The old tests were fairly bad, making it appear as though it supported pgsql + mysql, when in reality it only cared about pgsql (technically, I really only care about pgsql, but that doesn't mean that the test should ignore that if it's configured).